Update integration test jest.config.js files
Add separate configs for backend integration: mongodb, opensearch, postgres.
Update package.json to execute integration test for mongodb, opensearch, postgres.
