+++ .github/workflows/on-pullrequest.yml
@@ -156,7 +156,7 @@ jobs:
             },
             {
               type: "Integration",
-              command: "npm run test:integration -- --ci --config ./tests/config/integration/jest.config.js",
+              command: "npm run test:integration",
             },
           ]
