+++ Meadowlark-js/backends/meadowlark-opensearch-backend/test/setup/Setup.ts
@@ -2,13 +2,16 @@
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
-const openSearchEnvironmentTeardown = require('../../../backends/meadowlark-opensearch-backend/test/setup/OpenSearchSetupEnvironment');
 
-module.exports = async () => {
+const openSearchEnvironmentSetup = require('./OpenSearchSetupEnvironment');
+
+async function setupIntegrationTestEnvironment() {
   try {
     // Setup openSearch environment for integration tests.
-    await openSearchEnvironmentTeardown.teardownOpenSearch();
+    await openSearchEnvironmentSetup.setupOpenSearch();
   } catch (error) {
-    throw new Error(`Error Teardown: ${error}`);
+    throw new Error(`Error setting up integration test environment: ${error}`);
   }
-};
+}
+
+module.exports = async () => setupIntegrationTestEnvironment();
