+++ Meadowlark-js/backends/meadowlark-opensearch-backend/test/setup/Teardown.ts
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+const openSearchEnvironmentTeardown = require('./OpenSearchSetupEnvironment');
+
+module.exports = async () => {
+  try {
+    // Setup openSearch environment for integration tests.
+    await openSearchEnvironmentTeardown.teardownOpenSearch();
+  } catch (error) {
+    throw new Error(`Error Teardown: ${error}`);
+  }
+};
