+++ Meadowlark-js/backends/meadowlark-postgresql-backend/test/config/integration/jest.config.js
@@ -1,0 +1,17 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+const defaultConfig = require('../../../../../tests/config/jest.config');
+
+module.exports = {
+  displayName: 'Integration Tests: Postgresql',
+  ...defaultConfig,
+  testMatch: ['**/meadowlark-postgresql-backend/test/integration/**/*.(spec|test).[jt]s?(x)'],
+  coverageThreshold: {
+    global: {
+      branches: 52,
+      functions: 58,
+      lines: 60,
+      statements: 60,
+    },
+  },
+  workerIdleMemoryLimit: '200MB',
+};
