+++ Meadowlark-js/jest.config.js
@@ -1,8 +1,13 @@
+const { defaults: tsjPreset } = require('ts-jest/presets');
+
 module.exports = {
   projects: [
     '<rootDir>/tests/config/unit/jest.config.js',
     '<rootDir>/tests/config/integration/jest.config.js',
     '<rootDir>/tests/config/e2e/jest.config.js',
+    '<rootDir>/backends/meadowlark-mongodb-backend/test/config/integration/jest.config.js',
+    '<rootDir>/backends/meadowlark-opensearch-backend/test/config/integration/jest.config.js',
+    '<rootDir>/backends/meadowlark-postgresql-backend/test/config/integration/jest.config.js',
   ],
   coverageThreshold: {
     global: {
